Experiment with an OO interface
The class SPEWER is defined to hold the state previously in global
variables. The only global variable is a convenience on-demand singleton
when the API is called without first constructing a spewer.

The API is atypical in that there are no generic functions, since I
wanted to try to maintain backward compatibility with the old API. As
can be seen from the example functions, I think that has been
achieved (but I onlt tested write-strings).